added logic to write jsonencode object instead of escaped string

+++ tests.sh
@@ -1,10 +1,17 @@
+DATA_ACTIONS_PATH=$1
+
+if [ -z "$DATA_ACTIONS_PATH" ] ; then
+    echo "Please provide a path to the data action(s)"
+    exit 0
+fi
+
 MODULES_DIR="modules"
 
 if [ -d "$MODULES_DIR" ]; then
     rm -rf $MODULES_DIR
 fi
 
-python main.py testing
+python main.py --test $DATA_ACTIONS_PATH
 
 cd $MODULES_DIR
 
@@ -31,4 +38,4 @@ echo "\nAll tests passed."
 cd ..
 
 echo "\nResetting module directories..."
-./runner.sh
+./runner.sh $DATA_ACTIONS_PATH
